added kahoot-profane tool
In the future, it might be more convenient if this functionality were incorporated into the other tools. It is fairly easy to check if Kahoot would filter a nickname; then, and only then, we could apply the unicode hack (for *any* tool, not just kahoot-profane). For now, I want to have this be a separate tool so that people are more aware of the functionality.